requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
tensorflow==2.15.0
keras==3.12.0
numpy==1.24.3
Pillow==10.1.0
focal-loss==0.0.7
opencv-python==4.8.1.78
scipy==1.11.4
